Fix(#108): Criando nova função para verificação de datas.
